% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alicefunction.R
\name{alice_function}
\alias{alice_function}
\title{Alice Function}
\usage{
alice_function(a = 1, b = 0, c = 4, X = 1:12, Y = 42, Z = 21:30)
}
\arguments{
\item{a}{should be one real number}

\item{b}{a real number}

\item{c}{a real number}

\item{X}{a vector of real numbers (same length as Y and Z)}

\item{Y}{a vector of real numbers (same length as X and Z)}

\item{Z}{a vector of real numbers (same length as X and Y)}
}
\value{
This function returns the values computed by the function, and a graph of the polynomous function if 3 values or more are provided for X, Y and Z vectors
}
\description{
Alice Function
}
\examples{

}
