% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alicefunction.R
\name{alice_function}
\alias{alice_function}
\title{Alice Function}
\usage{
alice_function(a = 1, b = 0, c = 4, x = 1:12, y = 42, z = 21:30)
}
\arguments{
\item{a}{La reponse A}

\item{b}{La reponse B}

\item{c}{La reponse C}

\item{x}{Le parametre X}

\item{y}{Le parametre Y}

\item{z}{Le parametre Z}
}
\value{
This function returns the values computed by the function, or an error message
}
\description{
Alice Function
}
\examples{

}
